fix(ci): full-coverage maintainability gate — split auth_mounts.py overlay/Claude subsystem into auth_mounts_claude.py
src/awf/node/auth_mounts.py exceeded the 1500-line first-party limit. Extract
the generic overlayfs primitives (capability/filesystem probes, /proc/mounts
parsing, the OverlayMounter abstraction) and the full Claude auth subsystem
(host-signature hashing, shared-base build/reap, per-workspace overlay mount,
fallback reconciliation, teardown) into a new auth_mounts_claude.py sibling.
auth_mounts.py keeps service-mount resolution and the non-Claude provider
preps, and re-exports every moved public name so awf.node.auth_mounts.<name>
stays the stable import surface.

Keeping the overlay primitives and their Claude consumers in one module
preserves monkeypatch semantics: tests that patch _overlay_filesystem_available,
_has_cap_sys_admin, or _ensure_shared_claude_base now target the module that
defines them (auth_mounts_claude), where the consumers resolve them.
